fix(indent-hints): prevent nil function call (#13)
changed with nvim 0.11
"Query:iter_matches() correctly returns all matching nodes in a match instead of only the last node.
This means that the returned table maps capture IDs to a list of nodes that need to be iterated over.
For backwards compatibility, an option all=false (only return the last matching node)
is provided that will be removed in a future release."
https://neovim.io/doc/user/news-0.11.html